pyproject.toml
[project]
name = "friday"
version = "0.1.0"
description = "rag pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.34.2",
  "pymupdf==1.24.10",
  "sentence-transformers==4.1.0",
  "tqdm==4.66.5",
  "python-dotenv==1.0.1",
  "google-generativeai==0.8.2",
  "tenacity==9.0.0",
  "qdrant-client==1.11.0",
  "convex>=0.7.0",
  "fastapi-clerk-auth>=0.0.5",
]
